backend/requirements.txt
aiohttp==3.6.2
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==3.0.1
atb==0.1.5
attrs==24.2.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mido==1.3.3
multidict==4.7.6
nodeenv==1.9.1
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pretty_midi==0.2.10
propcache==0.2.1
psycopg2==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pygramtic==0.2.0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomlkit==0.13.2
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
yarl==1.18.3
